requirements.txt
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
beautifulsoup4==4.7.1
urllib3==1.24.3
oauth2client==4.1.3
python-dotenv==0.10.2
